+++ tests/test_commands.py
@@ -20,7 +20,7 @@ class CommandsTest(CliTestCase):
             # Example:
             destination = os.path.join(tmp_dir, "collection.json")
 
-            result = self.run_command(f"palsar2scansar create-collection {destination}")
+            result = self.run_command(f"palsar2-scansar create-collection {tmp_dir}")
 
             assert result.exit_code == 0, "\n{}".format(result.output)
 
@@ -28,7 +28,7 @@ class CommandsTest(CliTestCase):
             assert len(jsons) == 1
 
             collection = pystac.read_file(destination)
-            assert collection.id == "my-collection-id"
+            assert collection.id == "palsar2-scansar"
             # assert collection.other_attr...
 
             collection.validate()
@@ -38,18 +38,17 @@ class CommandsTest(CliTestCase):
             # Run your custom create-item command and validate
 
             # Example:
-            infile = "/path/to/asset.tif"
-            destination = os.path.join(tmp_dir, "item.json")
-            result = self.run_command(
-                f"palsar2scansar create-item {infile} {destination}"
-            )
+            item_id = "ALOS2437590500-220630_WWDR2.2GUA"
+            infile = f"tests/data-files/{item_id}_summary.xml"
+            destination = os.path.join(tmp_dir, f"{item_id}.json")
+            result = self.run_command(f"palsar2-scansar create-item {infile} {tmp_dir}")
             assert result.exit_code == 0, "\n{}".format(result.output)
 
             jsons = [p for p in os.listdir(tmp_dir) if p.endswith(".json")]
             assert len(jsons) == 1
 
             item = pystac.read_file(destination)
-            assert item.id == "my-item-id"
+            assert item.id == "ALOS2437590500-220630_WWDR2.2GUA"
             # assert item.other_attr...
 
             item.validate()